Actions for multiple items in list table

+++ app/Http/Controllers/Backend/GalleryController.php
@@ -34,6 +34,60 @@ class GalleryController extends Controller
     }
 
     /**
+     * Apply changes to multiple items
+     */
+    public function action(Request $request)
+    {
+
+        $action = $request->action;
+        $checked = $request->check;
+
+        switch ($action) {
+
+            case "publish":
+                foreach($checked as $k=>$v) {
+
+                    $gallery = Gallery::find($v);
+                    $gallery->status = 1;
+                    $gallery->save();
+                }
+                return redirect()->route('gallery')->with('status', 'Gallery are successfully set to active');
+                break;
+
+            case "unpublish":
+                foreach($checked as $k=>$v) {
+
+                    $gallery = Gallery::find($v);
+                    $gallery->status = 0;
+                    $gallery->save();
+                }
+                return redirect()->route('gallery')->with('status', 'Gallery are successfully set to inactive');
+                break;
+
+            default:
+                foreach($checked as $k=>$v) {
+
+                    $gallery = Gallery::find($v);
+
+                    $images = $gallery->images;
+                    
+                    if($images) {
+                        foreach($images as $i) {
+                            $i->delete();
+                            if(file_exists(public_path().'/uploads/images/gallery/'.$i->image)){
+                                @unlink(public_path().'/uploads/images/gallery/'.$i->image);
+                            }
+                        }
+                    }
+
+                    $gallery->delete();
+                }
+                $delete = true;
+                return redirect()->route('gallery', ['delete' => $delete])->with('status', 'Gallery are deleted successfully');
+        }
+    }
+
+    /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
@@ -101,12 +155,14 @@ class GalleryController extends Controller
     {
         $gallery_images = Gallery::find($id->id)->images;
 
-        foreach($gallery_images as $img) {
+        if($gallery_images) {
+            foreach($gallery_images as $img) {
 
-            $img->delete();
+                $img->delete();
 
-            if(file_exists(public_path().'/uploads/images/gallery/'.$img->image)){
-                @unlink(public_path().'/uploads/images/gallery/'.$img->image);
+                if(file_exists(public_path().'/uploads/images/gallery/'.$img->image)){
+                    @unlink(public_path().'/uploads/images/gallery/'.$img->image);
+                }
             }
         }
